feat(#826): delete core/types.py and core/operation_result.py shims — move callers to contracts
