+++ src/nexus/core/kernel_dispatch.py
@@ -36,6 +36,7 @@ from __future__ import annotations
 import logging
 from typing import Any
 
+from nexus.contracts.operation_result import OperationWarning
 from nexus.contracts.vfs_hooks import (
     DeleteHookContext,
     MkdirHookContext,
@@ -52,7 +53,6 @@ from nexus.contracts.vfs_hooks import (
     VFSWriteHook,
     WriteHookContext,
 )
-from nexus.core.operation_result import OperationWarning
 
 logger = logging.getLogger(__name__)
