--- src/nexus/core/types.py
@@ -1,10 +1,0 @@
-"""Backward-compat re-exports — canonical source is nexus.contracts.types."""
-
-from nexus.contracts.types import (
-    ContextIdentity,
-    OperationContext,
-    Permission,
-    extract_context_identity,
-)
-
-__all__ = ["ContextIdentity", "OperationContext", "Permission", "extract_context_identity"]
